refactor: move API functions into ConsensusApi
